Transition + Controller Software dev 1
I decided to make a seperate repository for this one, act a little more
professionally :smile:

The network is currently set up to utilize a file as a relay between the
server software and the controller software. I'll probably change that,
but I want it to get working in a basic form (i.e. establish simple
communications) first.